# Speed up cincoffset_impl()