--- a/target/cheri-common/cheri-lazy-capregs.h
+++ b/target/cheri-common/cheri-lazy-capregs.h
@@ -236,6 +236,37 @@ static inline void update_capreg(CPUArchState *env, unsigned regnum,
     cheri_log_instr_changed_gp_capreg(env, regnum, target);
 }
 
+
+/*
+ * This function can be called to avoid copying a full cap_register_t when
+ * updating a capability register inplace.
+ * Note: new_cursor must be representable given the bounds of source_cap;
+ * see addr_in_cap_bounds()/is_representable_cap_with_addr().
+ */
+static inline void update_capreg_cursor_from(CPUArchState *env, unsigned regnum,
+                                             const cap_register_t *source_cap,
+                                             unsigned source_regnum,
+                                             const target_ulong new_cursor)
+{
+    if (unlikely(regnum == NULL_CAPREG_INDEX)) {
+        return;
+    }
+    GPCapRegs *gpcrs = cheri_get_gpcrs(env);
+    cap_register_t *target = &gpcrs->decompressed[regnum];
+    cheri_debug_assert(get_capreg_state(gpcrs, source_regnum) ==
+                       CREG_FULLY_DECOMPRESSED);
+    cheri_debug_assert(is_representable_cap_with_addr(source_cap, new_cursor));
+    if (regnum != source_regnum) {
+        *target = *source_cap;
+        set_capreg_state(gpcrs, regnum, CREG_FULLY_DECOMPRESSED);
+    }
+    /* When updating in-place, we can avoid copying. */
+    target->_cr_cursor = new_cursor;
+    sanity_check_capreg(gpcrs, regnum);
+    rvfi_changed_capreg(env, regnum, target->_cr_cursor);
+    cheri_log_instr_changed_gp_capreg(env, regnum, target);
+}
+
 static inline void update_compressed_capreg(CPUArchState *env, unsigned regnum,
                                             target_ulong pesbt, bool tag,
                                             target_ulong cursor)